cdb_loader: moved functions to helper.cpp
